pyproject.toml
[project]
name = "notebookllm"
version = "0.1.0"
description = "NotebookLM-style RAG backend scaffold"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-docx>=1.1.0",
  "pdfplumber>=0.11.0",
  "pymupdf>=1.24.0",
  "beautifulsoup4>=4.12.0",
  "youtube-transcript-api>=0.6.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.27.0"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "fastapi*", "pydantic*"]
